Rename XP_Pen to XP
