--- OpenTabletDriver/Vendors/XP_Pen/XP_PenReportParser.cs
@@ -1,13 +1,0 @@
-using OpenTabletDriver.Tablet;
-using OpenTabletDriver.Plugin.Tablet;
-
-namespace OpenTabletDriver.Vendors.XP_Pen
-{
-    public class XP_PenReportParser : IReportParser<IDeviceReport>
-    {
-        public IDeviceReport Parse(byte[] data)
-        {
-            return (data[1] & 0xc0) == 0xc0 ? new XP_PenAuxReport(data) : new TabletReport(data);
-        }
-    }
-}
